+++ test_project/features/fixtures.py
@@ -1,6 +1,7 @@
 from behave import fixture
 
 from test_project.test_utils.dbt_test_utils import *
+from fixtures_sqlserver import *
 
 """
 The fixtures here are used to supply runtime metadata to tests, which are provided to the model generator.
@@ -115,64 +116,6 @@ def single_source_hub(context):
 
 
 @fixture
-def single_source_hub_sqlserver(context):
-    """
-    Define the structures and metadata to load single-source hubs
-    """
-
-    context.hashed_columns = {
-        "STG_CUSTOMER": {
-            "CUSTOMER_PK": "CUSTOMER_ID"
-        },
-        "STG_CUSTOMER_HASHLIST": {
-            "CUSTOMER_PK": ["CUSTOMER_ID", "CUSTOMER_NAME"]
-        }
-    }
-
-    context.vault_structure_columns = {
-        "HUB_CUSTOMER": {
-            "src_pk": "CUSTOMER_PK",
-            "src_nk": "CUSTOMER_ID",
-            "src_ldts": "LOAD_DATE",
-            "src_source": "SOURCE"
-        },
-        "HUB_CUSTOMER_SHA": {
-            "src_pk": "CUSTOMER_PK",
-            "src_nk": "CUSTOMER_ID",
-            "src_ldts": "LOAD_DATE",
-            "src_source": "SOURCE"
-        }
-    }
-
-    context.seed_config = {
-        "HUB_CUSTOMER": {
-            "+column_types": {
-                "CUSTOMER_PK": "BINARY(16)",
-                "CUSTOMER_ID": "VARCHAR(4)",
-                "LOAD_DATE": "DATETIME",
-                "SOURCE": "VARCHAR(4)"
-            }
-        },
-        "HUB_CUSTOMER_SHA": {
-            "+column_types": {
-                "CUSTOMER_PK": "BINARY(32)",
-                "CUSTOMER_ID": "VARCHAR(4)",
-                "LOAD_DATE": "DATETIME",
-                "SOURCE": "VARCHAR(4)"
-            }
-        },
-        "RAW_STAGE": {
-            "+column_types": {
-                "CUSTOMER_ID": "VARCHAR(4)",
-                "CUSTOMER_NAME": "VARCHAR(5)",
-                "LOAD_DATE": "DATETIME",
-                "SOURCE": "VARCHAR(4)"
-            }
-        }
-    }
-
-
-@fixture
 def multi_source_hub(context):
     """
     Define the structures and metadata to load multi-source hubs
@@ -243,82 +186,6 @@ def multi_source_hub(context):
                 "DISCOUNT": "NUMBER(38,2)",
                 "LOAD_DATE": "DATE",
                 "SOURCE": "VARCHAR"
-            }
-        }
-    }
-
-
-@fixture
-def multi_source_hub_sqlserver(context):
-    """
-    Define the structures and metadata to load multi-source hubs
-    """
-
-    context.hashed_columns = {
-        "STG_PARTS": {
-            "PART_PK": "PART_ID"
-        },
-        "STG_SUPPLIER": {
-            "PART_PK": "PART_ID",
-            "SUPPLIER_PK": "SUPPLIER_ID"
-        },
-        "STG_LINEITEM": {
-            "PART_PK": "PART_ID",
-            "SUPPLIER_PK": "SUPPLIER_ID",
-            "ORDER_PK": "ORDER_ID"
-        }
-    }
-
-    context.vault_structure_columns = {
-        "HUB": {
-            "src_pk": "PART_PK",
-            "src_nk": "PART_ID",
-            "src_ldts": "LOAD_DATE",
-            "src_source": "SOURCE"
-        }
-    }
-
-    context.seed_config = {
-        "HUB": {
-            "+column_types": {
-                "PART_PK": "BINARY(16)",
-                "PART_ID": "VARCHAR(4)",
-                "LOAD_DATE": "DATETIME",
-                "SOURCE": "VARCHAR(4)"
-            }
-        },
-        "RAW_STAGE_PARTS": {
-            "+column_types": {
-                "PART_ID": "VARCHAR(4)",
-                "PART_NAME": "VARCHAR(10)",
-                "PART_TYPE": "VARCHAR(10)",
-                "PART_SIZE": "VARCHAR(2)",
-                "PART_RETAILPRICE": "DECIMAL(11,2)",
-                "LOAD_DATE": "DATETIME",
-                "SOURCE": "VARCHAR(4)"
-            }
-        },
-        "RAW_STAGE_SUPPLIER": {
-            "+column_types": {
-                "PART_ID": "VARCHAR(4)",
-                "SUPPLIER_ID": "VARCHAR(2)",
-                "AVAILQTY": "INT",
-                "SUPPLYCOST": "DECIMAL(11,2)",
-                "LOAD_DATE": "DATETIME",
-                "SOURCE": "VARCHAR(4)"
-            }
-        },
-        "RAW_STAGE_LINEITEM": {
-            "+column_types": {
-                "ORDER_ID": "VARCHAR(5)",
-                "PART_ID": "VARCHAR(4)",
-                "SUPPLIER_ID": "VARCHAR(2)",
-                "LINENUMBER": "INT",
-                "QUANTITY": "INT",
-                "EXTENDED_PRICE": "DECIMAL(11,2)",
-                "DISCOUNT": "DECIMAL(11,2)",
-                "LOAD_DATE": "DATETIME",
-                "SOURCE": "VARCHAR(4)"
             }
         }
     }
